Film_BE/requirements_utf8.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.7
joblib==1.4.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==2.0.0
outcome==1.3.0.post0
pandas==2.2.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.14.0
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==7.0.0
pytest-mock==3.15.1
factory-boy==3.3.3
faker==12.0.1
freezegun==1.5.5
django-test-plus==2.3.0
model-bakery==1.20.5
mixer==7.2.2
google-generativeai>=0.3.0
openai>=1.0.0
